优化：将考试ID改为雪花ID

+++ Sourcecode/Song.Entities/ExamResults.cs
@@ -22,7 +22,7 @@ namespace Song.Entities {
     		
     		protected Int64 _Etp_Id;
     		
-    		protected Int32 _Exam_ID;
+    		protected Int64 _Exam_ID;
     		
     		protected String _Exam_Name;
     		
@@ -142,7 +142,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Exam_ID {
+    		public Int64 Exam_ID {
     			get {
     				return this._Exam_ID;
     			}
@@ -518,7 +518,7 @@ namespace Song.Entities {
     				this._Etp_Id = reader.GetInt64(_.Etp_Id);
     			}
     			if ((false == reader.IsDBNull(_.Exam_ID))) {
-    				this._Exam_ID = reader.GetInt32(_.Exam_ID);
+    				this._Exam_ID = reader.GetInt64(_.Exam_ID);
     			}
     			if ((false == reader.IsDBNull(_.Exam_Name))) {
     				this._Exam_Name = reader.GetString(_.Exam_Name);
@@ -654,7 +654,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Etp_Id = new WeiSha.Data.Field<ExamResults>("Etp_Id");
     			
     			/// <summary>
-    			/// 字段名：Exam_ID - 数据类型：Int32
+    			/// 字段名：Exam_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Exam_ID = new WeiSha.Data.Field<ExamResults>("Exam_ID");
